git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi-AOL@10822 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/src/rtcp/RTCPTimer.cpp
@@ -1,3 +1,20 @@
+// Copyright 2008 AOL LLC.
+// Licensed to SIPfoundry under a Contributor Agreement.
+//
+// This library is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// This library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this library; if not, write to the Free Software
+// Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301
+// USA. 
 //
 // Copyright (C) 2004-2006 SIPfoundry Inc.
 // Licensed by SIPfoundry under the LGPL license.
@@ -370,7 +387,7 @@ void  CRTCPTimer::ReportingAlarm(timer_t tTimer, int iArgument)
 #include <sys/time.h>
 #endif
 
-void CRTCPTimer::ReportingAlarm(const int userData, const int eventData)
+void CRTCPTimer::ReportingAlarm(const intptr_t userData, const intptr_t eventData)
 {
 #ifdef RTCP_LINUX_DEBUG
     struct timeval tv;
